+++ projects/ui/src/lib/components/splitter/index.ts
@@ -1,0 +1,3 @@
+export * from './kui-splitter.component';
+export * from './kui-splitter-orientation.type';
+export * from './kui-splitter-pane.component';